+++ src/main/kotlin/recharts/cartesian/area/AreaBaseValue.kt
@@ -4,9 +4,9 @@ import recharts.kotlin.Union
 
 @Suppress("ClassName")
 sealed class AreaBaseValue(name: String, ordinal: Int) : Union<AreaBaseValue>(name, ordinal) {
-    object dataMax : AreaBaseValue("dataMax", 0)
-    object dataMin : AreaBaseValue("dataMin", 1)
-    data class Number(val value: kotlin.Number) : AreaBaseValue("Number", 2)
+    data class Number(val value: kotlin.Number) : AreaBaseValue("Number", 0)
+    object dataMax : AreaBaseValue("dataMax", 1)
+    object dataMin : AreaBaseValue("dataMin", 2)
 
     companion object {
         fun valueOf(value: String): AreaBaseValue = when (value) {