+++ src/main/kotlin/recharts/kotlin/properties/AxisScaleOrNullProp.kt
@@ -1,13 +1,13 @@
 package recharts.kotlin.properties
 
+import recharts.kotlin.RechartsProps
 import recharts.util.types.AxisScale
-import recharts.util.types.BaseAxisProps
 import recharts.util.types.ScaleType
 import kotlin.properties.ReadWriteProperty
 import kotlin.reflect.KProperty
 
-class AxisScaleOrNullProp : ReadWriteProperty<BaseAxisProps, AxisScale?> {
-    override fun getValue(thisRef: BaseAxisProps, property: KProperty<*>) =
+class AxisScaleOrNullProp : ReadWriteProperty<RechartsProps, AxisScale?> {
+    override fun getValue(thisRef: RechartsProps, property: KProperty<*>) =
         when (val value = thisRef.asDynamic()[property.name]) {
             is Function<*> -> AxisScale.Function(value as Function<*>)
             is String -> try {
@@ -18,7 +18,7 @@ class AxisScaleOrNullProp : ReadWriteProperty<BaseAxisProps, AxisScale?> {
             else -> null
         }
 
-    override fun setValue(thisRef: BaseAxisProps, property: KProperty<*>, value: AxisScale?) {
+    override fun setValue(thisRef: RechartsProps, property: KProperty<*>, value: AxisScale?) {
         thisRef.asDynamic()[property.name] = when (value) {
             is AxisScale.Function -> value.value
             is AxisScale.ScaleType -> value.value.name