+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 
 ### Changed
 
+- Add missing properties (`interval`, `ticks`) to `XAxisProps`.
+- Add missing properties (`interval`, `ticks`) to `YAxisProps`.
+- Add missing properties (`scale`) to `ZAxisProps`.
 - Updated kotlin-react and kotlin-react-dom to `17.0.0-pre.129-kotlin-1.4.10`
 
 ## [v2.0.0-beta.7-4](https://gitlab.com/gm666q/kotlin-recharts/-/tree/v2.0.0-beta.7-4) - 2020-11-04